Q&A_Chatbot/requirements.txt
streamlit==1.33.0
python-dotenv==1.0.0
pinecone-client==3.0.0
sentence-transformers==2.5.1
langchain==0.1.4
langchain-groq==0.1.4
langchain-community==0.0.13
pdfplumber==0.10.3
